fix: 조회하기 구현완료 - 같은상품의 다른옵션상품이 담겨져있을경우 중복된 값들이 불러와지는 버그 해결 - 같은 상품의 다른옵션상품이 담겨져 있을경우 상품으로 묶어서 옵션의 정보들을 리스트로 제공
Resolves issue #9

+++ src/main/java/com/github/shoppingmallproject/repository/cart/CartEntity.java
@@ -20,7 +20,7 @@ public class CartEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "cart_id")
-    private Long cartId;
+    private Integer cartId;
 
     @ManyToOne
     @JoinColumn(name = "production_option_id", nullable = false)
